Addes piercing Bullets
Wall collision for piercing bullets is still buggy, but so far solved best as possible.
Collision after a wall resets after a timer, so hits behind a wall might not always trigger and longer walls might not be passable.
Maybe conditions inside the timer on when and what to reset and maybe to start itself again could help.